Fix windows build (size_t vs. int).
Sadly, Vector ctor takes an argument "int length".

TBR=dcarney@chromium.org, mstarzinger@chromium.org
BUG=

Review URL: https://codereview.chromium.org/334903002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@21844 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/ast-value-factory.h
@@ -233,11 +233,11 @@ class AstValueFactory {
         string_table_(AstString::Compare),
         zone_(zone),
         isolate_(NULL) {
-#define F(name, str) {      \
-      const char* data = str;          \
-      name##_string_ = GetOneByteString(                              \
+#define F(name, str) { \
+      const char* data = str; \
+      name##_string_ = GetOneByteString( \
           Vector<const uint8_t>(reinterpret_cast<const uint8_t*>(data), \
-                                strlen(data)));                         \
+                                static_cast<int>(strlen(data)))); \
     }
     STRING_CONSTANTS(F)
 #undef F
